+++ CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
-- Added `powerauth.service.rest.date.legacyFormatEnabled` property to restore the legacy `+00:00` timestamp offset on REST and callback responses for clients that cannot parse the Jackson 3 default `Z` designator [(#2436)](https://github.com/wultra/powerauth-server/issues/2436)
+### Fixed
+- Restored numeric (epoch-millisecond) serialization of operation callback timestamps, which regressed to ISO-8601 strings after the Jackson 3 upgrade and broke callback consumers that parse the timestamps as numbers [(#2436)](https://github.com/wultra/powerauth-server/issues/2436)
 
 ## [2.2.0] - 2026-07-20